Separate the tasks code from todo list into different file
- change onDragEnd function to have more readable code
- add provided.placeholder to get rid of warning in tasks
